issue-86: Additional filters for dynamic metrics
